Fix sampling of 0 probability tokens

+++ src/transformers/generation_utils.py
@@ -1559,7 +1559,7 @@ class GenerationMixin:
 
             # sample
             probs = F.softmax(next_token_scores, dim=-1)
-            next_tokens = torch.multinomial(probs, num_samples=1).squeeze(1)
+            next_tokens = torch.multinomial(probs.float(), num_samples=1).squeeze(1)
 
             # finished sentences should have their next token be a padding token
             if eos_token_id is not None:
@@ -2124,7 +2124,7 @@ class GenerationMixin:
 
             probs = F.softmax(next_token_scores, dim=-1)
 
-            next_tokens = torch.multinomial(probs, num_samples=2 * num_beams)
+            next_tokens = torch.multinomial(probs.float(), num_samples=2 * num_beams)
             next_token_scores = torch.gather(next_token_scores, -1, next_tokens)
 
             next_token_scores, _indices = torch.sort(next_token_scores, descending=True, dim=1)
